Plans 2023: purchases to data-store & use for pricing meta
