+++ packages/data-stores/package.json
@@ -34,6 +34,7 @@
 		"@automattic/calypso-analytics": "workspace:^",
 		"@automattic/format-currency": "workspace:^",
 		"@automattic/shopping-cart": "workspace:^",
+		"@automattic/js-utils": "workspace:^",
 		"@tanstack/react-query": "^4.36.1",
 		"@wordpress/api-fetch": "^6.44.0",
 		"@wordpress/data-controls": "^3.16.0",
