+++ packages/data-stores/src/purchases/hooks/use-site-purchase-by-id.ts
@@ -1,0 +1,17 @@
+import useSitePurchases from '../queries/use-site-purchases';
+import { Purchase } from '../types';
+
+interface Props {
+	purchaseId?: number | null;
+	siteId?: string | number | null;
+}
+
+const useSitePurchaseById = ( { siteId, purchaseId }: Props ): Purchase | undefined => {
+	const sitePurchases = useSitePurchases( { siteId } );
+
+	return sitePurchases?.data && typeof purchaseId === 'number'
+		? sitePurchases?.data[ purchaseId ]
+		: undefined;
+};
+
+export default useSitePurchaseById;
