+++ packages/data-stores/src/purchases/queries/lib/use-query-keys-factory.ts
@@ -1,0 +1,5 @@
+const useQueryKeysFactory = () => ( {
+	sitePurchases: ( siteId?: string | number | null ) => [ 'site-purchases', siteId ],
+} );
+
+export default useQueryKeysFactory;
